Creates card and buyer unit, validates valid card number.
Created uniqueness control for card and buyers records with their due exceptions and correct handling. Now the application is not able to register a buyer whose cpf is already registered and of a card whose identifying characteristics are already registered. Also created the wrong card number validation following the Luhn algorithm. Made change in the relationship of the payment and buy tables of many <-> one to one <-> one, making a buy only have one payment in the database